backend/pyproject.toml
[project]
name = "igreja-backend"
version = "1.0.0"
description = "Backend da Paróquia Santa Rosa de Viterbo"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "pyjwt>=2.10",
    "bcrypt>=4.2",
    "python-multipart>=0.0.18",
    "apscheduler>=3.10",
    "httpx>=0.28",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "taskipy>=1.14",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.taskipy.tasks]
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate"
createsuperuser = "python -m app.cli.createsuperuser"

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
